don't write to the .nyc_output directory
i was doing this because i wanted to get coverage reports from remote
runners and i still want to do so, but doing it this way was always
placeholder and wrong because it ends up merging with nyc's report and
so counts everything 2x. We should revisit this more intentionally at
some later point. But for now, we can support the latest version of nyc.