partially shore up code coverage drop
it's not going to be possible to test the code path where we don't
have a .stablerc file for the reason that we don't currently have
the infrastructure for testing alternative scenarios with stablercs.
This is because stable is "bootstrapped" in the sense that its test
suite is written against stable itself, and the the .stablerc tests
are written against the actual .stablerc files stable uses for its
own configuration!

This is suboptimal; there should be a way of testing different
scenarios outside of this project, but that's a larger topic that
needs another initiative to wrangle.